[#104] refactoring
