+++ youtora/scrape/scrp_mlgloss/scripts/prse_mlglossraw.py
@@ -1,0 +1,5 @@
+from ..models import MLGloss
+
+
+if __name__ == '__main__':
+    pass
