+++ youtora/scrape/scrp_mlgloss/scripts/scrp_mlglossraw.py
@@ -1,0 +1,5 @@
+#  script for scraping & and storing the data
+from ..models import MLGlossRaw
+
+if __name__ == '__main__':
+    pass
