+++ youtora/scrape/scrp_youtube/settings.py
@@ -1,8 +1,5 @@
 
 
-# scrapers settings
-# temporarily placed here
-# scraper.Caption.lang_code: (fields.E005) 'choices' must be an iterable containing
 # (actual value, human readable name) tuples.
 LANG_CODES_TO_COLLECT = (
     ('ko', 'korean'),
